Restore CI coverage for ESP32 targets

+++ src/channel_timeout.h
@@ -1,12 +1,13 @@
 #ifndef CHANNEL_TIMEOUT_H
 #define CHANNEL_TIMEOUT_H
 
+#include <cstdint>
+
 namespace channel_timeout {
 
-inline bool expired(unsigned long now, unsigned long lastActivity,
-                    unsigned long timeoutMs) {
+inline bool expired(uint32_t now, uint32_t lastActivity, uint32_t timeoutMs) {
   return timeoutMs > 0 && lastActivity > 0 &&
-         static_cast<unsigned long>(now - lastActivity) > timeoutMs;
+         static_cast<uint32_t>(now - lastActivity) > timeoutMs;
 }
 
 } // namespace channel_timeout
